chore(host): move inline tests to a separate file
The module-level test block in `host.rs` was extracted into its own file `host_tests.rs` to reduce the size of the main source file and keep test code separate from production logic. The change replaces the inline `mod tests` block with a `#[path = "host_tests.rs"] mod tests;` declaration, leaving only the import and the `host_of` function in the original file.

Auto-committed-on: dragonfly
